Add turn system
Need to redo the turn system again cuz it keeps
going into an infinite loop. I'll have to rethink
on how to move the game forward only when the player
has marked the grid. Maybe a boolean or counter
making sure to only move forward once the bool turns
true or the counter has increased and is greater
than the other persons? I'm not sure yet.